+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusCollectionManager.cs
@@ -172,8 +172,12 @@ internal sealed class PrometheusCollectionManager
 
         try
         {
+            var resourceTags = new PrometheusResourceTagCollection(this.exporter.Resource, this.exporter.AllowedResourceAttributesFilter);
+
             if (this.exporter.OpenMetricsRequested)
             {
+                cursor = PrometheusSerializer.WriteTargetInfo(this.buffer, cursor, this.exporter.Resource);
+
                 this.scopes.Clear();
 
                 foreach (var metric in metrics)
@@ -222,6 +226,7 @@ internal sealed class PrometheusCollectionManager
                             cursor,
                             metric,
                             this.GetPrometheusMetric(metric),
+                            resourceTags,
                             this.exporter.OpenMetricsRequested);
 
                         break;