+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusExporterOptions.cs
@@ -30,6 +30,11 @@ internal sealed class PrometheusExporterOptions
     }
 
     /// <summary>
+    /// Gets or sets the allowed resource attributes filter. Default value: No resource attributes allowed.
+    /// </summary>
+    public Predicate<string> AllowedResourceAttributesFilter { get; set; }
+
+    /// <summary>
     /// Gets or sets a value indicating whether addition of _total suffix for counter metric names is disabled. Default value: <see langword="false"/>.
     /// </summary>
     public bool DisableTotalNameSuffixForCounters { get; set; }