+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/Internal/PrometheusSerializerExt.cs
@@ -22,7 +22,7 @@ internal static partial class PrometheusSerializer
         return true;
     }
 
-    public static int WriteMetric(byte[] buffer, int cursor, Metric metric, PrometheusMetric prometheusMetric, bool openMetricsRequested = false)
+    public static int WriteMetric(byte[] buffer, int cursor, Metric metric, PrometheusMetric prometheusMetric, PrometheusResourceTagCollection resourceTags = default, bool openMetricsRequested = false)
     {
         cursor = WriteTypeMetadata(buffer, cursor, prometheusMetric);
         cursor = WriteUnitMetadata(buffer, cursor, prometheusMetric);
@@ -36,7 +36,7 @@ internal static partial class PrometheusSerializer
 
                 // Counter and Gauge
                 cursor = WriteMetricName(buffer, cursor, prometheusMetric);
-                cursor = WriteTags(buffer, cursor, metric, metricPoint.Tags);
+                cursor = WriteTags(buffer, cursor, metric, metricPoint.Tags, resourceTags);
 
                 buffer[cursor++] = unchecked((byte)' ');
 
@@ -87,7 +87,7 @@ internal static partial class PrometheusSerializer
 
                     cursor = WriteMetricName(buffer, cursor, prometheusMetric);
                     cursor = WriteAsciiStringNoEscape(buffer, cursor, "_bucket{");
-                    cursor = WriteTags(buffer, cursor, metric, tags, writeEnclosingBraces: false);
+                    cursor = WriteTags(buffer, cursor, metric, tags, resourceTags, writeEnclosingBraces: false);
 
                     cursor = WriteAsciiStringNoEscape(buffer, cursor, "le=\"");
 
@@ -113,7 +113,7 @@ internal static partial class PrometheusSerializer
                 // Histogram sum
                 cursor = WriteMetricName(buffer, cursor, prometheusMetric);
                 cursor = WriteAsciiStringNoEscape(buffer, cursor, "_sum");
-                cursor = WriteTags(buffer, cursor, metric, metricPoint.Tags);
+                cursor = WriteTags(buffer, cursor, metric, metricPoint.Tags, resourceTags);
 
                 buffer[cursor++] = unchecked((byte)' ');
 
@@ -127,7 +127,7 @@ internal static partial class PrometheusSerializer
                 // Histogram count
                 cursor = WriteMetricName(buffer, cursor, prometheusMetric);
                 cursor = WriteAsciiStringNoEscape(buffer, cursor, "_count");
-                cursor = WriteTags(buffer, cursor, metric, metricPoint.Tags);
+                cursor = WriteTags(buffer, cursor, metric, metricPoint.Tags, resourceTags);
 
                 buffer[cursor++] = unchecked((byte)' ');