+++ src/OpenTelemetry.Exporter.Prometheus.HttpListener/PrometheusHttpListenerOptions.cs
@@ -41,4 +41,9 @@ public class PrometheusHttpListenerOptions
             this.uriPrefixes = value;
         }
     }
+
+    /// <summary>
+    /// Gets or sets the allowed resource attributes filter. Default value: No resource attributes allowed.
+    /// </summary>
+    public Predicate<string> AllowedResourceAttributesFilter { get; set; }
 }