+++ test/OpenTelemetry.Exporter.Prometheus.HttpListener.Tests/PrometheusHttpListenerTests.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Net.Http;
 #endif
 using OpenTelemetry.Metrics;
+using OpenTelemetry.Resources;
 using OpenTelemetry.Tests;
 using Xunit;
 
@@ -84,6 +85,12 @@ public class PrometheusHttpListenerTests
     }
 
     [Fact]
+    public async Task PrometheusExporterHttpServerIntegration_AddResourceAttributeAsTag()
+    {
+        await this.RunPrometheusExporterHttpServerIntegrationTest(addServiceNameResourceTag: true);
+    }
+
+    [Fact]
     public void PrometheusHttpListenerThrowsOnStart()
     {
         Random random = new Random();
@@ -154,7 +161,10 @@ public class PrometheusHttpListenerTests
             });
     }
 
-    private async Task RunPrometheusExporterHttpServerIntegrationTest(bool skipMetrics = false, string acceptHeader = "application/openmetrics-text")
+    private async Task RunPrometheusExporterHttpServerIntegrationTest(
+        bool skipMetrics = false,
+        string acceptHeader = "application/openmetrics-text",
+        bool addServiceNameResourceTag = false)
     {
         var requestOpenMetrics = acceptHeader.StartsWith("application/openmetrics-text");
 
@@ -175,9 +185,15 @@ public class PrometheusHttpListenerTests
             {
                 provider = Sdk.CreateMeterProviderBuilder()
                     .AddMeter(meter.Name)
+                    .ConfigureResource(x => x.Clear().AddService("my_service", serviceInstanceId: "id1"))
                     .AddPrometheusHttpListener(options =>
                     {
                         options.UriPrefixes = new string[] { address };
+
+                        if (addServiceNameResourceTag)
+                        {
+                            options.AllowedResourceAttributesFilter = s => s == "service.name";
+                        }
                     })
                     .Build();
 
@@ -232,15 +248,22 @@ public class PrometheusHttpListenerTests
 
             var content = await response.Content.ReadAsStringAsync();
 
+            var resourceTagAttributes = addServiceNameResourceTag
+                ? "service_name='my_service',"
+                : string.Empty;
+
             var expected = requestOpenMetrics
-                ? "# TYPE otel_scope_info info\n"
+                ? "# TYPE target info\n"
+                  + "# HELP target Target metadata\n"
+                  + "target_info{service_name='my_service',service_instance_id='id1'} 1\n"
+                  + "# TYPE otel_scope_info info\n"
                   + "# HELP otel_scope_info Scope metadata\n"
                   + $"otel_scope_info{{otel_scope_name='{MeterName}'}} 1\n"
                   + "# TYPE counter_double_total counter\n"
-                  + $"counter_double_total{{otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+\\.\\d{{3}})\n"
+                  + $"counter_double_total{{{resourceTagAttributes}otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+\\.\\d{{3}})\n"
                   + "# EOF\n"
                 : "# TYPE counter_double_total counter\n"
-                  + $"counter_double_total{{otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+)\n"
+                  + $"counter_double_total{{{resourceTagAttributes}otel_scope_name='{MeterName}',otel_scope_version='{MeterVersion}',key1='value1',key2='value2'}} 101.17 (\\d+)\n"
                   + "# EOF\n";
 
             Assert.Matches(("^" + expected + "$").Replace('\'', '"'), content);